backend/pyproject.toml
[project]
name = "mockup-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0,<1.0",
    "uvicorn[standard]>=0.30.0,<1.0",
    "anthropic>=0.42.0,<1.0",
    "openai>=1.50.0,<2.0",
    "asyncpg>=0.30.0,<1.0",
    "sqlalchemy[asyncio]>=2.0.30,<3.0",
    "motor>=3.6.0,<4.0",
    "certifi>=2024.2.2",
    "pydantic-settings>=2.5.0,<3.0",
    "httpx>=0.27.0,<1.0",
    "python-dotenv>=1.0.0,<2.0",
    "python-jose[cryptography]>=3.3.0,<4.0",
    "bcrypt>=4.1.0,<5.0",
    "tiktoken>=0.7.0",
    # ── BYO Material Pipeline ──
    "PyMuPDF>=1.24.0",
    "yt-dlp>=2024.0",
    "youtube-transcript-api>=0.6",
    "elevenlabs>=1.0",
    "google-cloud-storage>=2.18",
    "Pillow>=10.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]
